+++ VisLang/DataNodes.cs/Arrays/ArrayGetElementAtNode.cs
@@ -13,16 +13,16 @@ public class ArrayGetElementAtNode : DataNode
     {
     }
 
-    public int DefaultIndex = 0;
+    public long DefaultIndex = 0;
 
     public Value? GetInputArray(NodeContext? context) => Inputs.ElementAtOrDefault(0)?.GetValue(context);
-    public int GetInputIndex(NodeContext? context) => Inputs.ElementAtOrDefault(1)?.GetValue(context)?.AsInteger() ?? DefaultIndex;
+    public long GetInputIndex(NodeContext? context) => Inputs.ElementAtOrDefault(1)?.GetValue(context)?.AsInteger() ?? DefaultIndex;
 
     public override Value? GetValue(NodeContext? context = null)
     {
         if (GetInputArray(context)?.Data is List<Value> arr)
         {
-            return arr[GetInputIndex(context)];
+            return arr[(int)GetInputIndex(context)];
         }
         return null;
     }
